Marks: Hide process cleanup in explain mode
Do not show killing processes at the end of a test when in explain mode.

git-svn-id: https://svn.geocomp.uq.edu.au/svn/jeathesis@154 a4bfb809-27fc-4812-b90d-531e79a1b4ec

+++ src/libmarks/py_src/case.py
@@ -203,6 +203,10 @@ class TestCase(object):
 
     def _cleanup_processes(self):
         """Attempt to kill all processes started within a test"""
+        if self.option('explain'):
+            # Do not cleanup, as no processes are running.
+            return
+
         for p in self._processes:
             try:
                 p.kill()
@@ -287,7 +291,8 @@ class TestCase(object):
             msg = "Timeout occurred"
 
         # Kill process, to ensure it is not left around.
-        process.kill()
+        if not self.option('explain'):
+            process.kill()
 
         raise self.failure_exception(msg)
